review/requirements.txt
grpcio
grpcio-tools
loguru
PyYAML
firebase-admin
pydantic
kafka-python
beanie